requirements.txt
anyio==3.6.1
click==8.1.3
fastapi==0.85.0
h11==0.14.0
idna==3.4
kafka-python==2.0.2
pydantic==1.10.2
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.3.0
uvicorn==0.18.3
